Segundo Commit
Se agrega Modelo Pedido

+++ demo/src/main/java/com/example/demo/service/UsuarioServiceImpl.java
@@ -15,6 +15,7 @@ public class UsuarioServiceImpl implements UsuarioService {
     @Autowired
     private UsuarioRepository usuarioRepository;
 
+
     @Override
     public List<Usuario> getAllUsuarios(){
         return usuarioRepository.findAll();
@@ -42,8 +43,14 @@ public class UsuarioServiceImpl implements UsuarioService {
     }
 
     @Override
-    public void deleteUsuario(Integer id){
-        usuarioRepository.deleteById(id);
+    public boolean deleteUsuario(Integer id){
+        if (usuarioRepository.existsById(id)) {
+            usuarioRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
-    
+
+   
 }
